Switch from time to chrono
This is a breaking change. The time module is deprecated, and
suggests switching over to the chrono library, which is more
actively maintained. One thing to note is that chrono does not
support parsing with `%Z`:

https://github.com/chronotope/chrono/issues/38

However, this actually isn't that bad. According to RFC 2616, the only
supported timezone is "GMT":

http://tools.ietf.org/html/rfc2616#section-3.3.1

Furthermore, @lifthrasiir points out that most implementations of `%Z`
is actually not widely supported. On my mac, it claims to only support
parsing the local timezone, and GMT, so I believe it's safe to hardcode
the timezone.

This also fixes a bug in an example, where October 21, 2017 would be a
Saturday. However, Wednesday 21 October, 2015 is Back to the Future
day, so this switches to using that date instead.