Switch max-age over to u64.
This is a breaking change. In #85, it was suggested that we remove
the public dependency on `time::Duration` since that crate is deprecated.
According to the [Cookie spec](https://tools.ietf.org/html/rfc6265#section-5.2.2),
`Max-Age` is in seconds that is either zero or always positive. Negative
numbers are always parsed to the earliest representable date and time,
which would be 0.