Bump `aes-gcm` to v0.9.1 and enable `force-soft`; MSRV 1.41+
I released a new version of `aes-gcm` with a `force-soft` feature:

https://github.com/RustCrypto/AEADs/pull/306

This disables autodetection of AES-NI, falling back on a portable
constant-time software implementation.

However, in this configuration `aes-gcm` still supports an MSRV of
1.41+, as opposed to 1.49+ required by the autodetection feature.

A comment in Cargo.toml notes `force-soft` should be removed when MSRV
is bumped to at least 1.49.