Revert b62b4e: strip quotes from cookie value.
Some applications expect quotes to be preserved in a cookie's value, as
indicated by the RFC and expected by Servo, while others expect quotes to be
trimmed, as implemented by libraries like Django and Express. This commit
reverts a previous commit that always trimmed quotes and leaves the decision to
the user by introducing a new `value_trimmed()` getter method. The `value()`
method is reverted to the previous behavior, and the `value_trimmed()` strips
quotes.