Implement implicit cookie creation.
This commit Implements and leverages new `T: Into<Cookie>` `impl`s to simplify
the creation of and reference to cookies across the crate. Specifically, it adds
implementations of `Into<Cookie>` for string-like types (`&str`, `String`,
`Cow<str>`), tuples of string-like types `(name string, value string)`, and
`CookieBuilder`. The former implementations create a cookie with a name
corresponding to the string and an empty value. The tuple implementation creates
a cookie with the given name and value strings. The `CookieBuilder`
implementation simply unwraps the in-process cookie.

Functions and methods that previously accepted a `Cookie` now accept any `T:
Into<Cookie>`. This particularly affects the `CookieJar` API, which now allows
simpler addition and removal of cookies:

  * `jar.add(("foo", "bar"));`
  * `jar.add(Cookie::build(("foo", "bar")).path("/"));`
  * `jar.remove("foo");`
  * `jar.remove(Cookie::build("foo").path("/"));`

Additionally, as noted above, the `Into<Cookie>` `impl` for `CookieBuilder`
obviates most explicit calls to `CookieBuilder::finish()` (now `build()`).

Resolves #199.