chore: add rust-version to Cargo.toml
This helps the MSRV aware resolver pick a version compatible with
the project. This version is also displayed by crates.io, which is
nice. It may also obsolete the versions table in README.md given
that crates.io shows the `rust-version` field in the versions
history page, for example: https://crates.io/crates/tokio/versions